api: sanitize job ids from api calls
The job_id returned by sourcehut (or local test stub) should be a
decimal, "all characters in the string are decimal characters and there
is at least one character". Also it should match the image/package in
the database of which we assume the job is building. Checking this makes
sense in general, but I'm adding it here to prepare using the job id in
a temp dir path in a future patch.