Implement staging repositories
Improve the workflow for testing service packs (and possibly other
feature branches such as the lomiri one). It is now possible to push a
so-called staging branch in the format <original-branch>_staging_<name>,
for example: "v23.06_staging_sp1". BPO will figure out that this is a
staging branch without additional configuration. It will sync packages
(create hardlinks) from the original branch that have the same version
on the staging branch, and build the missing packages on the staging
branch, and then publish a full repository that can be used instead of
the repository for the orignal branch.

Syncing with the original branch is done whenever preparing to build the
next package, so it is possible to later on rebase the feature branch on
the original branch and force push, BPO will figure out the differences
again and build only that.

To delete the staging repository, delete the branch.