build_image: install lang and musl-locales as part of every image (MR 77)
Localization and internationalization are important topics, specially
for those people that are not devs used to work with English and C
locale by default. Since bpo-installed images are mostly aimed at less
technical users that will not use pmbootstrap, installing those packages
here would greatly simplify the process that those users need to follow
to get the system in their language. It would no longer be necessary to
open a terminal and install some packages.