sourcehut: put apk cache into tmpfs
Reduce the total disk space needed to build the installer images by
storing the apk cache in tmpfs. When building a pinephone plasma mobile
installer image, this results in using ~900 MiB less disk space.

This gives a bit of relief to not hit the space limit until we can
overhaul how the installer images are created with ondev2.

Related: pmaports issue 1623