images: add pretty_name for release name (MR 80)
The pretty name can be used by installers to add some extra information,
that it can show the state of a release, like "v23.06 (work in
progress)" or "v22.06 (end of life)".

Also let's add more tests for a v23.06 (stable) and v23.12 (WIP) version
so all the code paths are tested.