Don't output EOL packages and images
* Don't return packages and images from EOL releases
* Order packages by branch-arch-pkgname (right now they don't have a
  defined order and get returned in the order they were written into the
  database)
* Add tests

Adding a script to scrub EOL entries from the DB could be done in a
separate commit. Even if we don't want to list all of them on the status
page, it is useful to keep the information in the database a bit longer,
as the db is used to generate the html pages for all releases (EOL
releases may be still around on disk, we don't delete them right when a
release goes EOL).