data/templates/index: link to badge name
When the UI gets updated, the current badge (building, up-to-date,
failed) gets copied to /badge.svg. Adjust the template to not show the
badge that was copied to /badge.svg, but rather show the badge by its
name. This avoids caching problems, which would lead to displaying the
wrong badge for some time after it has been copied over.