bpo.repo.has_unfinished_builds: check for queued
Queued builds are unfinished too. Not having this there has lead to
assuming the WIP repository is done when in reality it wasn't yet, and
later failing in a sanity check for the WIP repo with e.g.:

RuntimeError: Found package in database, but not in WIP or final repository: phosh-0.30.0-r2.apk

One way to trigger this was apparently having a package depend on
itself.