bpo.db.Package: ignore depend on itself
Currently phosh on v23.06_staging_sp1 depends on itself, due to a
parsing bug (it depends on phosh-keyboard, which is not built yet, and
pmbootstrap figures it's probably a subpackage of phosh whereas in
reality it gets provided by squeekboard). Ignore depends of packages
that point to themselves. If they should really be the reason why we
can't built a package, it will fail when actually trying to build it.