Revert "bpo.db.Package: ignore depend on itself"
Actually this doesn't really help. If pmbootstrap and bpo conclude that
a package depends on itself, it is not useful to start the job and let
it fail again because pmbootstrap figures the package depends on itself
and that dependency is missing. I wasn't sure that would happen but now
I am, so revert the patch.

This reverts commit f0b871ecee53e94d4b03e238031a05d663ae3e2b.