bpo.repo.staging: copy file instead of hardlink
With the new server, we have the staging repository on a different
storage volume. Creating a hardlink to the WIP repository does not work
anymore, so copy the file for now.

The logic for hardlinks was slightly broken anyway, once files were
copied out of the WIP repository, they would get fully copied instead of
preserving the hardlink as Clayton found in issue 128.

Let's rather exclude the staging repo dir from rsync. Using a bit more
space on our server is fine... or actually with btrfs + copy on write it
might not even use more space with copying once the file gets published
from WIP repo to staging repo proper.