repo_bootstrap: don't clear entire wip repo dir
When combining a staging repository with repo_bootstrap, as I'm
currently doing for systemd, the staging repository may already have
synced packages in the wip repository when repo_bootstrap is running. Do
not remove all packages from the WIP repo, instead only remove the
packages that are out-of-sync with the database.

The previous code had an error in the glob line (wrong argument count).