db: make two columns not unique
Usually this is not the way to do it, and migration should be done
instead. However, setting columns to non-unique is not trivial with
sqlite, it only works when removing the table and adding it again.

Since we only have one instance of this, I decided to simply adjust the
bpo.db on the server this time.

Why the unique constrains are removed:
* Package.job_id: with repo_bootstrap multiple packages get generated
  with the same job. This was not possible before.
* RepoBootstrap.dir_name: this was a copy-paste mistake, it must be
  possible to have multiple RepoBootstrap entries for different branches
  and architectures with the same dir_name (typically / or /systemd).

Related: https://stackoverflow.com/questions/42013265/remove-unique-constraint-on-a-column-in-sqlite-database