bpo.repo.bootstrap: add log msg for published
Bring back the log msg for published [repo_bootstrap]. I had this in a
development version, and then removed it again because at this point,
the actual list of packages that were published gets printed. But after
deploying, I realized that with many other log messages appearing it is
useful to explicitly log that [repo_bootstrap] is finished.