test/test_zz_50_repo_bootstrap: fix assert
With the recent patch that replaced print() with logging, an assert has
appeared for this test even though it has passed.

Let the fake_step4 function, which is monkeypatched as
bpo.jobs.build_package, return True when it is done. This way
bpo.repo.build_arch_branch assumes that the package build has been
started successfully and will not attempt to trigger another package
build (which would then cause fake_step4 to get called again, and cause
the assert since it gets called in an unexpected order).

The test passed anyway, because bpo_test.stop_server() was called in
fake_step4, to stop the test successfully.

While at it, remove a commented out monkeypatch to set fake_step4, and
put the proper call in that position to make the flow a bit clearer.