bpo/jobs/build_image: no 'continue' for !installer
Do not continue in the loop if we are not building an installer image,
and don't modify task_name so it could be used afterwards.

I've thought about doing android-recovery-zip installer in that loop
after installer image. But space-wise it makes more sense to do it after
images are compressed. Still, keep this commit because it makes the loop
slightly cleaner.