repo_bootstrap: set dir_name to splitrepo (MR 119)
Instead of having dir_name always set to "/", iterate over the packages
in the current branch and set dir_name to the splitrepo the package is
in, if that repo needs a bootstrap.

In practice this means, the dir_name will have "systemd" as value
instead of "/" from now on (and we could even bootstrap more split
repos in the future if needed).

This patch is needed for split repository logic, so we know that the
bootstrap needs to be done on the split "systemd" repository and not
the main one.