bpo/job_services/sourcehut: fix using pmbv3 config
Adjust the code path that sets "is_default_channel = False" in the
pmbootstrap config to use the correct path for the pmbv3 config, if
using pmbv3.

This piece of code ensures that pmbootstrap won't try to change the
channel to edge during "pmbootstrap init" (by changing the pmaports
branch if channel != edge). It didn't work with pmbv3 yet - we only
didn't notice it because we only used pmbv3 so far for "master" and
"master_staging_systemd".