templates: fix linking to non-existing wiki pages
Link to a search page instead of the assumed wiki title. With the
"&title=Special:Search" argument, it works the same as if using the
search function on the wiki page itsself: if a page with the exact name
exists, the user will be redirected there. Otherwise a search will be
shown, which usually has the page we mean on top (e.g. search for
"qcom-msm8953" has "Generic MSM8953 (qcom-msm8953)" on top).

Fixes: #122