db: add column package.splitrepo (MR 119)
In the current bpo code, the term "repo" refers to wip/final/symlink
repos. To avoid confusion when introducing the new split
(normal/systemd) repo, call it "splitrepo".

Previous code has never read from the "repo" column, and before
introducing split repos we want the default value to be NULL for all
packages. I've considered removing the "repo" column, but that would
make the migration code more complex. Just keep it for now, using a bit
more space for that is fine.