testdata/depends_systemd: add pkg to normal repo (MR 119)
Add another package for the regular repository in testdata. Splitrepos
are only built after the main repository has been built and published.

Having a package with the same name in the main and a split repo needs
to be supported, so testing this here is useful as well.

Related: issue 144