test/bpo_test.raise_exception: set level to error (MR 119)
Use log level "error" instead of "critical", as pytest is configured in
this repo to show critical errors during the log output. Triggering
exceptions in testcases is expected (to test such exceptions), and if we
should get an unexpected exception then the test already aborts. With
log level "error" we don't get cluttered output in regular test runs.