jobs/sign_index: use set_repos_task
Replace the logic of setting PMB_APK_FORCE_MISSING_REPOSITORIES=1 if the
main repository APKINDEX does not exist with generating a proper
set_repos_task that sets mirrors.pmaports and mirrors.systemd, depending
on the APKINDEX for each existing. The previous method did not work if
the main repo APKINDEX existed, but the systemd one did not.

Fix for:
  [07:33:24] WARNING: file not found: http://mirror.postmarketos.org/postmarketos/extra-repos/systemd/v25.06/x86_64/APKINDEX.tar.gz
  [07:33:24] NOTE: check the [mirrors] section in 'pmbootstrap config'
  [07:33:24] ERROR: getting APKINDEX from binary package mirror failed!

Related: https://builds.sr.ht/~postmarketos/job/1503227